Fix version of grunt-contrib-coffee
Previously, npm was installing the latest version of this package.
In the latest version, the Coffeescript code compiled didn't transform
the ES6 class syntax to ES5, eg: it kept `class Filter` in the code.

Uglify doesn't parse ES6 code and it crashed.

This commit changes the major version of the grunt-contrib-coffee package
to 1. This version transform the ES6 class syntax to ES5 and Uglify can do
its job.